Handler and client streaming tracing + refactor (#41)
## Changes:
- Refactor, 
- Implement streaming handler tracing
- Implement streaming client tracing
- Add AttributeFilter
- Remove `Tracer()` method and `traceprovider` on config and just
initialize a single tracer for opentelemetry-connect-go
- Consolidate metrics and span recording together; don't 

### fix existing bugs:
- `success` is recorded instead of any error in `StreamingClient`
- `requests` and `responses` metrics are currently recorded in
`streamingState`, but it has been moved to the interceptors because
`requests` are `send` events for clients and `receive` events for
handlers (and vice versa for `responses`).
- both `instr.requestsPerRPC.Record` and `instr.responsesPerRPC.Record`
are currently called within `streamingstate.receive` which is incorrect
because then any `receive` event records both events when it should be
one. At least one of these calls should be in `send` depending on what
type of handler it is.
- inject trace context from unary client; I don't see any reason why a
span trace can't be handed to a unary call.
- Don't record events when err == io.EOF
- Set span events and attributes at the end of the interceptor (this
helps with readability quite a lot)

### Attribute filter
There is now an option to filter based on attribute.KeyValue + Request. 
This is enabled by passing in an option
`otelconnect.WithAttributeFilter` that takes a `Request` and
`attribute.KeyValue`. Example of filtering out `NetPeerPort` for server
metrics only:
```go
otelconnect.WithAttributeFilter(func(request *otelconnect.Request, value attribute.KeyValue) bool {
		if !request.Spec.IsClient && (value.Key == semconv.NetPeerPortKey || value.Key == semconv.NetPeerIPKey) {
			return false
		}
		return true
	})
```