Remove WithTelemetry Option in favor of NewInterceptor (#60)
The only use case we have of using this package is using a number of
Interceptors instead of just the `WithTelemetry` option. It's
anticipated that if a project is using connect-opentelemetry-go the
chances are that it's probably also using other interceptors. In that
case it makes sense to expose a `New` function that returns an `type
interceptor struct` that can be plugged into the existing interceptor
plumbing.

The only difference this makes to existing projects that are using
`WithTelemetry` is that they now need to use
`connect.WithInterceptor(otelconnect.New())`

Fixes: https://github.com/bufbuild/connect-opentelemetry-go/issues/21