Update test protos for PingStream method (#140)
Test protos were using `CumSum` as a `Ping` method that returned twice
and all other methods were unused. Regenerate the `PingService` with a
`PingStream` methods that returns a full duplex method of Ping. Size is
removed from the constant 2 to be calculated with `proto.Size()`
(response sizes are the same as request, so response isn't calculated).

Testing changes:
- `CumSum` == `PingStream` but always returns a response for a request,
not two responses for each request
- `failPing` returns an error, not a message than an error.

No behaviour should be changed.