Eagerly create instruments to fail fast on initialization (#141)
Change in API on creating `Interceptor`. Instruments are created eagerly
and any errors fail on init. We therefore return the error as part of
NewInterceptor, instead of deferring to first use and then failing user
RPCs (which is a serious no-no for instrumentation and goes against
recommended behavior:
https://opentelemetry.io/docs/specs/otel/error-handling/#basic-error-handling-principles
This is a backwards-incompatible change, but should be trivial for users
to correct on upgrade.