Remove *Request to fix streaming client context (#153)
Context in client streams doesn't include the span due to the `*Request`
requiring information only included once the connection is created. This
PR removes the `*Request` in favor of simply providing the `connect.Spec`
information. This allows for correctly initializing the span so its included in
the context passed to the machinery to issue the call.

Backward-incompatible changes:
* Removes exported `*Request` object (breaking change)
* Changes option filters to use `connect.Spec` instead of `*Request`
(breaking change)