Add option to force uniform output (#196)
This adds the `WithRPCSystem` option, which forces the output to either
use Connect or gRPC semantic conventions. When the option is not
provided, which conventions are used depends on the actual wire protocol
of the request being measured (same as before this change).

This also removes the "grpc_web" value for the "rpc.system" attribute
since that is not specified as a valid value anywhere in [OpenTelemetry
RPC semantic
conventions](https://opentelemetry.io/docs/specs/semconv/rpc/). The gRPC
semantic conventions suggest that it should always be "grpc" --
"grpc-web" is just a transport-level dialect and should not change
telemetry.